requirements.txt
streamlit
pdfplumber
pytesseract
Pillow
pandas
PyMuPDF